# Support fetching storable vectors + use them in benchmark