@fkm3 2016-12-13
I couldn't find a way to deal with overlapping instances, so I just listed all of the instances manually. Tensorflow only supports a small set of element types, so this seems reasonable to me. Let me know what you think.