@fkm3 2016-12-13
Half baked idea: Instead of

```haskell
class TensorDataType s a where
    ...

instance TensorDataType V.Vector Float
```

we could make it

```haskell
class TensorDataType a where
    ...

instance TensorDataType (V.Vector Float)
```

then, we could drop `Scalar` and allow scalars to be fetched directly

```haskell
instance TensorDataType Float
```